+++ Core/Atlas.lua
@@ -301,14 +301,12 @@ local function bossButtonUpdate(button, encounterID, instanceID, b_iconImage, mo
 			if (iconImage) then
 				SetPortraitTextureFromCreatureDisplayID(button.bgImage, displayInfo)
 			end
-			-- if (only if you're currently in this instance) then
-			-- will have to fire this function when you zone in or out, it doesn't seem to fire now
-			-- also todo: set defeated text in tooltip on button in tooltip on boss list, and change the boss pic in the list to an x
-			local complete = C_EncounterJournal.IsEncounterComplete(encounterID);
-			button.DefeatedOpacity:SetShown(complete);
-			button.DefeatedOverlay:SetShown(complete);
-			button.bgImage:SetDesaturation(complete and 0.7 or 0);
-			--end
+			if button.DefeatedOverlay then
+				local complete = C_EncounterJournal.IsEncounterComplete(encounterID);
+				button.DefeatedOpacity:SetShown(complete);
+				button.DefeatedOverlay:SetShown(complete);
+				button.bgImage:SetDesaturation(complete and 0.7 or 0);
+			end
 		end
 	end
 end
@@ -766,17 +764,18 @@ function addon:MapAddNPCButton()
 			if (info_x == nil) then info_x = -18; end
 			if (info_y == nil) then info_y = -18; end
 
-			if (WoWRetail and info_id < 10000 and profile.options.frames.showBossPotrait) then
+			if ((WoWRetail or WoWClassic) and info_id < 10000 and profile.options.frames.showBossPotrait) then
 				bossbutton = _G["AtlasMapBossButton"..bossindex]
+				local template = WoWRetail and "AtlasFrameBossButtonTemplate" or "AtlasFrameBossButtonTemplateClassic"
 				if (not bossbutton) then
-					bossbutton = CreateFrame("Button", "AtlasMapBossButton"..bossindex, AtlasFrame, "AtlasFrameBossButtonTemplate")
+					bossbutton = CreateFrame("Button", "AtlasMapBossButton"..bossindex, AtlasFrame, template)
 				end
 				bossButtonCleanUp(bossbutton)
 				bossButtonUpdate(bossbutton, info_id, base.JournalInstanceID, true)
 
 				bossbuttonS = _G["AtlasMapBossButtonS"..bossindexS]
 				if (not bossbuttonS) then
-					bossbuttonS = CreateFrame("Button", "AtlasMapBossButtonS"..bossindexS, AtlasFrameSmall, "AtlasFrameBossButtonTemplate")
+					bossbuttonS = CreateFrame("Button", "AtlasMapBossButtonS"..bossindexS, AtlasFrameSmall, template)
 				end
 				bossButtonCleanUp(bossbuttonS)
 				bossButtonUpdate(bossbuttonS, info_id, base.JournalInstanceID, true)
@@ -1058,7 +1057,7 @@ function Atlas_MapRefresh(mapID)
 	AtlasText_LevelRange_Text:SetText(tLR)
 
 	-- Map's Recommended Level Range
-	if (WoWRetail) then
+	if (WoWRetail or WoWClassic) then
 		local tRLR = ""
 		if (minRecLevel or minRecLevelH or minRecLevelM) then
 			local tmp_RLR = L["ATLAS_STRING_RECLEVELRANGE"]..L["Colon"]